Move cli to sub-library cabal-fmt-internal
